+++ db-queue-scheduler-core/src/main/java/ru/yoomoney/tech/dbqueue/scheduler/internal/queue/ScheduledTaskQueueConsumer.java
@@ -10,7 +10,6 @@ import ru.yoomoney.tech.dbqueue.api.impl.NoopPayloadTransformer;
 import ru.yoomoney.tech.dbqueue.scheduler.config.ScheduledTaskLifecycleListener;
 import ru.yoomoney.tech.dbqueue.scheduler.internal.ScheduledTaskDefinition;
 import ru.yoomoney.tech.dbqueue.scheduler.internal.db.ScheduledTaskQueueDao;
-import ru.yoomoney.tech.dbqueue.scheduler.internal.db.ScheduledTaskRecord;
 import ru.yoomoney.tech.dbqueue.scheduler.internal.schedule.ScheduledTaskExecutionContext;
 import ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTask;
 import ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTaskContext;
@@ -35,6 +34,8 @@ import static java.util.Objects.requireNonNull;
  */
 class ScheduledTaskQueueConsumer implements QueueConsumer<String> {
     private static final Logger log = LoggerFactory.getLogger(ScheduledTaskQueueConsumer.class);
+
+    private static final Duration MIN_HEARTBEAT_INTERVAL = Duration.ofSeconds(10L);
 
     private final QueueConfig queueConfig;
     private final ScheduledTaskDefinition scheduledTaskDefinition;
@@ -76,49 +77,80 @@ class ScheduledTaskQueueConsumer implements QueueConsumer<String> {
         scheduledTaskLifecycleListener.started(scheduledTaskDefinition.getIdentity(), scheduledTaskContext);
         log.debug("execute(): scheduledTaskIdentity={}, task={}", scheduledTaskDefinition.getIdentity(), task);
 
-        ScheduledTaskRecord queueTask = scheduledTaskQueueDao.findQueueTask(queueConfig.getLocation().getQueueId()).orElseThrow();
-
         long start = clock.millis();
         ScheduledTaskExecutionContext internalContext = new ScheduledTaskExecutionContext();
+        internalContext.setAttemptsCount(task.getAttemptsCount());
         ScheduledTaskExecutionResult executionResult = executeTask(scheduledTaskContext, internalContext);
-
         long processingTaskTime = clock.millis() - start;
 
-        if (executionResult.getType() == ScheduledTaskExecutionResult.Type.ERROR) {
-            log.debug("task executed: executionResult={}, nextExecutionTime={}", executionResult, queueTask.getNextProcessAt());
-            executionResult.getNextExecutionTime().ifPresent(nextExecutionTime ->
-                    scheduledTaskQueueDao.updateNextProcessDate(queueConfig.getLocation().getQueueId(), nextExecutionTime));
-            scheduledTaskLifecycleListener.finished(scheduledTaskDefinition.getIdentity(), scheduledTaskContext, executionResult,
-                    queueTask.getNextProcessAt(), processingTaskTime);
-            return TaskExecutionResult.fail();
-        }
-
         Instant nextExecutionTime = executionResult.getNextExecutionTime().orElseGet(() ->
-                        scheduledTaskDefinition.getNextExecutionTimeProvider().getNextExecutionTime(internalContext));
+                scheduledTaskDefinition.getNextExecutionTimeProvider().getNextExecutionTime(internalContext));
 
         log.debug("task executed: executionResult={}, nextExecutionTime={}", executionResult, nextExecutionTime);
+
         scheduledTaskLifecycleListener.finished(scheduledTaskDefinition.getIdentity(), scheduledTaskContext, executionResult,
                 nextExecutionTime, processingTaskTime);
+        if (executionResult.getType() == ScheduledTaskExecutionResult.Type.ERROR) {
+            scheduledTaskQueueDao.updateNextProcessDate(queueConfig.getLocation().getQueueId(), nextExecutionTime);
+            return TaskExecutionResult.fail();
+        }
         return TaskExecutionResult.reenqueue(Duration.between(clock.instant(), nextExecutionTime));
     }
 
     private ScheduledTaskExecutionResult executeTask(ScheduledTaskContext scheduledTaskContext,
                                                      ScheduledTaskExecutionContext internalContext) {
 
+        ScheduledTaskExecutionResult result;
         internalContext.setLastExecutionStartTime(clock.instant());
+
+        HeartbeatAgent heartbeatAgent = createHeartbeatAgent(internalContext);
         try {
-            ScheduledTaskExecutionResult result = scheduledTaskDefinition.getScheduledTask().execute(scheduledTaskContext);
+            heartbeatAgent.start();
+            result = scheduledTaskDefinition.getScheduledTask().execute(scheduledTaskContext);
             if (result.getState().isPresent()) {
                 scheduledTaskQueueDao.updatePayload(queueConfig.getLocation().getQueueId(), result.getState().orElseThrow());
             }
-            return result;
         } catch (RuntimeException ex) {
             scheduledTaskLifecycleListener.crashed(scheduledTaskDefinition.getIdentity(), scheduledTaskContext, ex);
             log.debug("failed to execute scheduled task: scheduledTask={}", scheduledTaskDefinition, ex);
-            return ScheduledTaskExecutionResult.error();
+            result = ScheduledTaskExecutionResult.error();
         } finally {
             internalContext.setLastExecutionFinishTime(clock.instant());
+            heartbeatAgent.stop();
         }
+        internalContext.setExecutionResultType(result.getType());
+        return result;
+    }
+
+    /**
+     * Creates heartbeat agent that helps to postpone next execution date-time of the task in case of time-consuming
+     * execution of the current one. That helps to prevent concurrent execution of the same task.
+     *
+     * @param internalContext internal context of a current execution
+     * @return prepared heartbeat agent
+     */
+    private HeartbeatAgent createHeartbeatAgent(ScheduledTaskExecutionContext internalContext) {
+        ScheduledTaskExecutionContext failInternalContext = internalContext.copy();
+        failInternalContext.setExecutionResultType(ScheduledTaskExecutionResult.Type.ERROR);
+
+        Duration precomputeNextExecutionDelay = Duration.between(
+                clock.instant(),
+                scheduledTaskDefinition.getNextExecutionTimeProvider().getNextExecutionTime(failInternalContext)
+        );
+
+        Duration heartbeatInterval = MIN_HEARTBEAT_INTERVAL.compareTo(precomputeNextExecutionDelay.dividedBy(2L)) > 0
+                ? MIN_HEARTBEAT_INTERVAL
+                : precomputeNextExecutionDelay.dividedBy(2L);
+
+        return new HeartbeatAgent(
+                scheduledTaskDefinition.getIdentity().asString(),
+                heartbeatInterval,
+                () -> shiftNextExecutionTime(heartbeatInterval.multipliedBy(2L))
+        );
+    }
+
+    private void shiftNextExecutionTime(Duration interval) {
+        scheduledTaskQueueDao.updateNextProcessDate(queueConfig.getLocation().getQueueId(), clock.instant().plus(interval));
     }
 
     @Nonnull